test(test262): inject assert harness for Date slice
Add a minimal shared Test262 assert module and inject const { assert } = require('assert'); for the targeted Date 15.9.1.15-1 fixture so the port no longer needs an inlined copied harness block.
